+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.51.3
+- [🚀] 增加`Rect.scale`扩展;
+
 ## 0.51.2
 - [🚀] `FutureX`增加`thenIf`和`thenIfNot`扩展;
 - [🚀] `SafeAreaConfig`增加`minimum`参数;